size64

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	char c;
 	float f;
 
-	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(i));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(I));
-	printf("Size of long long int: %lu byte(s)\n", (unsigned long)sizeof(L));
-	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+	printf("Size of a char: %zu byte(s)\n", sizeof(c));
+	printf("Size of an int: %zu byte(s)\n", sizeof(i));
+	printf("Size of a long int: %zu byte(s)\n", sizeof(I));
+	printf("Size of long long int: %zu byte(s)\n", sizeof(L));
+	printf("Size of a float: %zu byte(s)\n", sizeof(f));
 	return (0);
 }
